added a custom and default skeleton file to separate those routes
