+++ reaxpress/helpers/skeleton/custom.jsx
@@ -1,0 +1,1 @@
+export default [];
