+++ reaxpress/helpers/skeleton/defaults.jsx
@@ -1,0 +1,11 @@
+export default [
+  {
+    "key": "index",
+    "routes": [
+      {
+        "path": "/",
+        "component": "Index"
+      }
+    ]
+  }
+];
