+++ reaxpress/helpers/skeleton/read.jsx
@@ -1,4 +1,4 @@
-import skeleton from '../../skeleton';
+import skeleton from './custom';
 
 const read = () => skeleton;
