+++ reaxpress/helpers/skeleton/remove.jsx
@@ -1,4 +1,4 @@
-import skeleton from '../../skeleton';
+import skeleton from './custom';
 import write from './write';
 
 const remove = (name) => {
