+++ reaxpress/skeleton.js
@@ -1,1 +1,3 @@
-module.exports = [];
+import { defaults, custom } from './helpers/skeleton';
+
+module.exports = [...defaults, ...custom];
